reformatted hadamard for functionals
